+++ lib/glimmer/swt/custom/shape/focus.rb
@@ -40,6 +40,7 @@ module Glimmer
           def include?(x, y)
             x.between?(self.x, self.x + width) && y.between?(self.y, self.y + height)
           end
+          alias contain? include?
           
         end
       end